requirements.txt
agno
google-genai
